# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WSC Access Workbench API
  version: 2.2.0
  description: |
    wsc-contracts@2.2.0 — 接入端工作台 V1.4（座序图 L2 Top5、持久化用户、我的数据产品；SNAP-WSC-005）。
    统一响应包装：{ code, message, data, correlationId }；成功时 code="0"。
    RBAC：产品增/改/导入仅 PROVIDER；ADMIN 保留分类维护与用户管理；目录维护 ADMIN+PROVIDER（PROVIDER 仅 create_by=本人）。
    GET /catalog/products?mine=true 仅返回 create_by=当前用户的产品。
    GET /catalog/l2-distribution 返回二级目录产品计数（座序图）。
    ADMIN /admin/users CRUD 持久化 sys_user。
    产品编码格式（DEC-WSC-001）：`{DOMAIN}-{FEATURE}-{NNNN}`。
    认证：服务端会话（Cookie）；写 API 以服务端会话角色为准。
    （HOTFIX TASK-WSC-607：目录维护对 PROVIDER 开放本人产品；契约版本号未 bump，仍为 2.2.0。）
servers:
  - url: /api/v1
    description: V1 API base
...
security:
  - SessionCookie: []
paths:
...
  /catalog/categories:
    get:
      tags: [catalog]
      summary: 行业分类树
      operationId: listCategories
      responses:
        "200":
          description: 一二三级分类（空间/行业/子类）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCategoryList"
    post:
      tags: [catalog]
      summary: 新增行业分类（管理员）
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryWrite"
      responses:
        "200":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCategory"
        "403":
          $ref: "#/components/responses/Forbidden"
  /catalog/categories/{categoryId}:
    put:
      tags: [catalog]
      summary: 更新行业分类（管理员）
      operationId: updateCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryWrite"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCategory"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [catalog]
      summary: 删除行业分类（管理员；有挂载产品则禁止）
      operationId: deleteCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: 已删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseEmpty"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 分类下仍有产品
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                hasProducts:
                  value:
                    code: ERR_CATEGORY_HAS_PRODUCTS
                    message: 分类下仍有挂载产品，禁止删除
                    data:
                      reason: 分类下仍有 3 个产品
                      productCount: 3
                    correlationId: "corr-cat-del-1"
...
  /catalog/products:
    get:
      tags: [catalog]
      summary: 产品列表（筛选/搜索）
      operationId: listProducts
      parameters:
        - name: l1CategoryId
          in: query
          schema: { type: string }
        - name: l2CategoryId
          in: query
          schema: { type: string }
          description: 行业（二级）；可选
        - name: l3CategoryId
          in: query
          schema: { type: string }
          description: 子类（三级）；可选
        - name: industryCategory
          in: query
          schema: { type: string }
          description: GB/T 4754 门类中文原文（产品字段 industryCategory）；可选；与 L2/L3 树筛选独立
        - name: dataSource
          in: query
          schema: { $ref: "#/components/schemas/DataSource" }
        - name: productType
          in: query
          schema: { $ref: "#/components/schemas/ProductType" }
        - name: involvesPublicData
          in: query
          schema: { type: boolean }
        - name: deliveryMethod
          in: query
          schema: { $ref: "#/components/schemas/DeliveryMethod" }
        - name: q
          in: query
          description: 产品名/编码包含匹配（大小写不敏感）
          schema: { type: string }
        - name: mine
          in: query
          description: true 时仅返回 create_by=当前用户的产品（我的数据产品）
          schema: { type: boolean, default: false }
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: 分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProductPage"
    post:
      tags: [catalog]
      summary: 新增数据产品（仅 PROVIDER；ADMIN/USER → 403）
      description: |
        产品写仅 PROVIDER；ADMIN 与 USER 调用返回 ERR_FORBIDDEN（403）。
        归属 create_by 绑定当前会话 userId。
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWrite"
            examples:
              dataset:
                summary: 数据集
                value:
                  productCode: MED-EMR-0002
                  productName: 电子病历数据集
                  productType: DATASET
                  industryCategory: 卫生和社会工作
                  l3CategoryId: "cat-l3-1"
                  typeSpecific:
                    dataset:
                      recordCount: 1000
              other:
                summary: 其他数据产品（允许 contentDescription）
                value:
                  productCode: GEN-MISC-0001
                  productName: 其他产品示例
                  productType: OTHER
                  industryCategory: 建筑业
                  typeSpecific:
                    other:
                      contentDescription: 其他数据产品内容说明
                      timeRange: "2026/01/01 -"
                      regionScope: 全国
              api:
                summary: 数据接口（OpenAPI 形状）
                value:
                  productCode: MED-API-0002
                  productName: 核验接口
                  productType: API
                  industryCategory: 信息传输、软件和信息技术服务业
                  typeSpecific:
                    api:
                      swaggerFileContent: "openapi: 3.0.3\ninfo:\n  title: demo\n  version: 1.0.0\npaths: {}\n"
                      endpoints:
                        - id: ep-1
                          method: POST
                          path: /enterprise/security/verify
                          summary: 企业安全信息核验
      responses:
        "200":
          description: 已创建并产生第 1 版上链
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProduct"
        "400":
          description: 格式非法或校验失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                badCode:
                  value:
                    code: ERR_PRODUCT_CODE_FORMAT
                    message: 产品编码格式非法，须符合 {DOMAIN}-{FEATURE}-{NNNN}
                    data: null
                    correlationId: "corr-prod-fmt-1"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 编码冲突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                conflict:
                  value:
                    code: ERR_PRODUCT_CODE_CONFLICT
                    message: 产品编码已存在
                    data: null
                    correlationId: "corr-prod-dup-1"
  /catalog/products/{productId}:
    get:
      tags: [catalog]
      summary: 产品详情
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: 详情（分组字段）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProduct"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [catalog]
      summary: 编辑数据产品（仅 PROVIDER 本人；ADMIN/USER/异主 → 403）
      description: |
        产品写仅 PROVIDER 且 create_by=当前用户；空归属/异主/ADMIN/USER → ERR_FORBIDDEN（403）。
      operationId: updateProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWrite"
      responses:
        "200":
          description: 已更新并追加上链版本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProduct"
        "400":
          description: 编码格式或 OQ-004
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                badCode:
                  value:
                    code: ERR_PRODUCT_CODE_FORMAT
                    message: 产品编码格式非法，须符合 {DOMAIN}-{FEATURE}-{NNNN}
                    data: null
                    correlationId: "corr-prod-fmt-2"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 编码冲突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseError"
              examples:
                conflict:
                  value:
                    code: ERR_PRODUCT_CODE_CONFLICT
                    message: 产品编码已存在
                    data: null
                    correlationId: "corr-prod-dup-2"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: WSC_SESSION
      description: |
        服务端会话 Cookie。登录后写入；登出立即失效。
        角色绑定在会话中；写 API 以服务端会话角色为准。
        角色/权限变更后须吊销或强制刷新旧会话。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-JWT
      description: |
        可选等价方案：Bearer 令牌必须映射可吊销的服务端会话表。
        非「纯无状态 JWT」；吊销后写 API 必须拒绝。
  parameters:
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string }
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string }
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
          examples:
            forbidden:
              value:
                code: ERR_FORBIDDEN
                message: 当前角色无权执行该写操作
                data: null
                correlationId: "corr-403"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponseError"
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, data, correlationId]
      properties:
        code:
          type: string
          description: 成功为 "0"；业务错误为 ERR_* 或 HTTP 语义码字符串
        message:
          type: string
        data:
          nullable: true
        correlationId:
          type: string
    ApiResponseEmpty:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              nullable: true
    ApiResponseError:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            code:
              type: string
              enum:
                - ERR_PRODUCT_CODE_FORMAT
                - ERR_PRODUCT_CODE_CONFLICT
                - ERR_CATEGORY_HAS_PRODUCTS
                - ERR_CATEGORY_LEAF_REQUIRED
                - ERR_MAINTENANCE_FORBIDDEN
                - ERR_IMPORT_FILE_TOO_LARGE
                - ERR_IMPORT_FORMAT_INVALID
                - ERR_IMPORT_TEMPLATE_UNSUPPORTED
                - ERR_IMPORT_REPORT_NOT_FOUND
                - ERR_FORBIDDEN
                - ERR_ROLE_SWITCH_DISABLED
                - ERR_USER_USERNAME_CONFLICT
                - ERR_USER_NOT_FOUND
                - "401"
                - "404"
            data:
              nullable: true
              description: ERR_CATEGORY_HAS_PRODUCTS 时含 reason、productCount
...
    ProductType:
      type: string
      enum: [DATASET, REPORT, API, OTHER]
      description: |
        DATASET=数据集, REPORT=数据报告, API=数据接口, OTHER=其他数据产品。
        OTHER：允许 typeSpecific.other（含 contentDescription 及共用时间/地域）；
        不得再要求 OTHER 必须为空对象。
    DataSource:
      type: string
      enum: [SELF_PRODUCED, PUBLIC_COLLECT, AGREEMENT, DERIVED]
    DeliveryMethod:
      type: string
      enum: [API, FILE, SANDBOX, PRIVACY_COMPUTE]
    UpdateFrequency:
      type: string
      enum: [REALTIME, DAILY, WEEKLY, MONTHLY, YEARLY, ON_DEMAND, NO_UPDATE]
      description: |
        REALTIME=实时, DAILY=每日, WEEKLY=每周, MONTHLY=每月, YEARLY=每年,
        ON_DEMAND=按需, NO_UPDATE=不更新（V1.3 / OQ-V13-008）。
...
    Category:
      type: object
      required: [id, name, level]
      properties:
        id: { type: string }
        name: { type: string }
        level:
          type: string
          enum: [L1, L2, L3]
          description: L1=空间 L2=行业 L3=子类
        parentId:
          type: string
          nullable: true
        pathLabels:
          type: array
          items: { type: string }
          description: 可选；自根到本节点的名称路径
    CategoryWrite:
      type: object
      required: [name, level]
      properties:
        name: { type: string }
        level:
          type: string
          enum: [L1, L2, L3]
        parentId:
          type: string
          nullable: true
          description: L2 父为 L1；L3 父为 L2；L1 省略
    ApiResponseCategory:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Category"
    ApiResponseCategoryList:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/Category"
    ProductCode:
      type: string
      pattern: '^[A-Z0-9]+-[A-Z0-9]+-[0-9]{4,}$'
      description: DEC-WSC-001 `{DOMAIN}-{FEATURE}-{NNNN}`
      example: MED-EMR-0002
    ApiEndpointParameter:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        required: { type: boolean }
        description: { type: string }
    ApiEndpointResponse:
      type: object
      properties:
        code: { type: string }
        description: { type: string }
    ApiEndpoint:
      type: object
      properties:
        id: { type: string }
        method: { type: string, description: "GET|POST|PUT|DELETE|PATCH 等" }
        path: { type: string }
        summary: { type: string }
        description: { type: string }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpointParameter" }
        responses:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpointResponse" }
        responseBodySchema: { type: string }
        requestBodySchema: { type: string }
    TypeSpecificApi:
      type: object
      description: |
        数据接口专属。归档真源 swaggerFileContent；派生存贮 endpoints[]。
        写冲突（二者均非空且不一致）时以客户端 endpoints 为准，禁止服务端静默重算覆盖。
        兼容旧字段 endpoint（非唯一真源）。
      properties:
        swaggerFileContent: { type: string, description: 原始 OpenAPI/Swagger 文本；可空 }
        endpoints:
          type: array
          items: { $ref: "#/components/schemas/ApiEndpoint" }
        fieldDescription: { type: string }
        dataSample: { type: string }
        timeRange: { type: string }
        regionScope: { type: string }
        endpoint: { type: string, description: 兼容旧字段；可读；写非唯一真源 }
    TypeSpecificDataset:
      type: object
      properties:
        timeRange: { type: string }
        regionScope: { type: string }
        dataScale: { type: string }
        dataForm: { type: string, description: 图片/文本/视频/音频/表格/其他 }
        fieldDescription: { type: string }
        dataSample: { type: string }
    TypeSpecificReportOrOther:
      type: object
      description: report 与 other 均允许 contentDescription + 共用时间/地域
      properties:
        timeRange: { type: string }
        regionScope: { type: string }
        contentDescription: { type: string }
    TypeSpecificFields:
      type: object
      description: |
        按 productType 择一填充 api / dataset / report / other。
        OTHER 允许 other.contentDescription 及共用时间/地域（修订 OQ-004）；
        不再要求 OTHER 必须为空对象。
      properties:
        dataset:
          $ref: "#/components/schemas/TypeSpecificDataset"
        report:
          $ref: "#/components/schemas/TypeSpecificReportOrOther"
        api:
          $ref: "#/components/schemas/TypeSpecificApi"
        other:
          $ref: "#/components/schemas/TypeSpecificReportOrOther"
    IndustryCategory:
      type: string
      description: |
        GB/T 4754—2017 门类中文原文（产品字段，非目录树节点；DEC-WSC-006）。
        枚举对齐 v0729 模板「行业分类」列 dataValidation。
      enum:
        - 农、林、牧、渔业
        - 采矿业
        - 制造业
        - 电力、热力、燃气及水生产和供应业
        - 建筑业
        - 批发和零售业
        - 交通运输、仓储和邮政业
        - 住宿和餐饮业
        - 信息传输、软件和信息技术服务业
        - 金融业
        - 房地产业
        - 租赁和商务服务业
        - 科学研究和技术服务业
        - 水利、环境和公共设施管理业
        - 居民服务、修理和其他服务业
        - 教育
        - 卫生和社会工作
        - 文化、体育和娱乐业
        - 公共管理、社会保障和社会组织
        - 国际组织
    ProductWrite:
      type: object
      required: [productCode, productName, productType, industryCategory]
      properties:
        productCode:
          $ref: "#/components/schemas/ProductCode"
        productName: { type: string }
        productType:
          $ref: "#/components/schemas/ProductType"
        industryCategory:
          $ref: "#/components/schemas/IndustryCategory"
        l3CategoryId:
          type: string
          nullable: true
          description: 可选目录树三级挂载；非行业分类语义；可空（方案 A）
        businessCategory: { type: string }
        businessSubCategory: { type: string }
        dataSource: { $ref: "#/components/schemas/DataSource" }
        updateFrequency: { $ref: "#/components/schemas/UpdateFrequency" }
        deliveryMethod: { $ref: "#/components/schemas/DeliveryMethod" }
        involvesPersonalInfo: { type: boolean }
        involvesPublicData: { type: boolean }
        billingMethod: { type: string }
        price: { type: string }
        supplierName: { type: string }
        supplierCreditCode:
          type: string
          description: 供应商信用代码（敏感；日志禁止明文）
        propertyRightsType: { type: string }
        tags:
          type: array
          items: { type: string }
        summary: { type: string }
        scenario: { type: string }
        typeSpecific:
          $ref: "#/components/schemas/TypeSpecificFields"
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductWrite"
        - type: object
          required: [id, chainCount]
          properties:
            id: { type: string }
            chainCount: { type: integer }
            categoryPath: { type: string }
            latestVersionNo: { type: integer }
            createBy:
              type: string
              nullable: true
              description: 创建者 userId；mine=true 过滤依据；空/缺失不可被 PROVIDER 冒领
            updatedAt: { type: string, format: date-time }
    ProductPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiResponseProduct:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Product"
    ApiResponseProductPage:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ProductPage"
...